Extract reception, aspects, radicality and serialization modules
