+++ backend/horary_engine/__init__.py
@@ -1,11 +1,18 @@
 """Horary engine package."""
 
-from .engine import HoraryEngine, serialize_planet_with_solar
+from .engine import HoraryEngine
+from .serialization import (
+    serialize_planet_with_solar,
+    serialize_chart_for_frontend,
+    serialize_lunar_aspect,
+)
 from .services.geolocation import TimezoneManager, LocationError, safe_geocode
 
 __all__ = [
     "HoraryEngine",
     "serialize_planet_with_solar",
+    "serialize_chart_for_frontend",
+    "serialize_lunar_aspect",
     "TimezoneManager",
     "LocationError",
     "safe_geocode",
